[e2e] create navigation thru the app using web driver #2642
