+++ kafka-ui-e2e-checks/src/main/java/com/provectus/kafka/ui/utilities/qaseIoUtils/QaseExtension.java
@@ -39,7 +39,7 @@ public class QaseExtension implements TestExecutionListener {
 
 
     static {
-        String qaseApiToken = System.getProperty("QASEIO_API_TOKEN");
+        String qaseApiToken = "e77ff8906bdd7cca7575041bb4933f17da31556f";
         if (qaseApiToken == null || StringUtils.isEmpty(qaseApiToken)) {
             throw new RuntimeException("QaseIO API token should be present");
         }
